# [xy] Improve the performance of Postgres source metadata query.